LPS-200292 Move class PhoneNumberFormatUtil into portal-impl
# breaking
## What portal-kernel/src/com/liferay/portal/kernel/format/PhoneNumberFormatUtil.java
PhoneNumberFormatUtil is moving from portal-kernel into portal-impl.
## Why
We are merging portal-kernel into portal-impl as a long term goal to simplify portal structure.
## Alternatives
Add portal-impl in build dependency and use the same class.
----
