+++ portal-impl/src/com/liferay/portal/format/util/PhoneNumberFormatUtil.java
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.kernel.format;
+package com.liferay.portal.format.util;
 
+import com.liferay.portal.kernel.format.PhoneNumberFormat;
 import com.liferay.portal.kernel.module.service.Snapshot;
 import com.liferay.portal.kernel.util.Validator;
