+++ portal-impl/src/com/liferay/portal/format/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
